modify group calendar for weekday status
